pyproject.toml
[project]
name = "photo-doc-data-embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "numpy>=2.2.0,<2.4.0",
    "pillow>=12.0.0",
    "qdrant-client>=1.16.2",
    "transformers>=4.57.3",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "streamlit>=1.52.2",
    "opencv-python>=4.11.0.86",
    "scikit-learn>=1.8.0",
    "hdbscan>=0.8.41",
]

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
